Mode prefix caching benchmark to a separate file.
